For inference servers, setup fake tokenizer so know model's maximum token length
